Replace mules with a cargo mech and rollerskates on Meta (#56489)
This PR replaces the two mulebots in the warehouse on Meta with a cargo
variant of the ripley and two pairs of rollerskates. The cargo variant
of the Ripley has only two maximum equipment slots, starts with 50
integrity, and a smaller power cell. A recharging station for the
hauler is available in cargo. (Apparently the station doesn't need a
computer to function).

Mulebots have historically been incredibly poor performing. They're
slow, unresponsive, clunky, and when they do work properly, they only
work in batches of one crate and require you to radio in that something
has been delivered. More often than not, since you aren't dealing with
people, your deliveries are left forgotten under plastic flaps in a
department. Most of the time, people either ignore the Mulebots and
leave them to gather dust in the corner or hack them to use them as
murder machines instead of for their primary purpose.

I've been hearing that people feel cargo techs don't have has much to
do during the shift as people would like, so I thought it would be a
good idea to give them some tools that will help them do their job in a
way that interacts with the crew. The rollerskates are a great way to
get around if you're hauling singular crates (and are kind of funny if
you think of the techies as busboys). The cargo hauler should make it
much easier to make batch deliveries personally and I've attempted to
minimize any abuses people may have with it by limiting its equipment
slots.

If the changes are well received, I think it would be cool to
eventually replace the cargo hauler with a forklift vehicle.