Plants don't get their positions ruined when dropped (#58806)
Tiny atomic one line change that makes plants stay in the same place when you drop them.

What's the point of being able to hide in a plant if anyone with basic object permanence can see that a plant that used to be offset off the tile is now centered?
Also I was tired of seeing all the plants in the bar turn really ugly and offset the second Pun Pun gets to them with no way of centering them again.